Rename context files
Doing this in a separate commit so the diff is less confusing
